pyproject.toml
[build-system]
requires = ["hatchling >=0.25"]
build-backend = "hatchling.build"

[project]
name = "tomorrowcities"
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]
dependencies = [
    "solara", 
    "geopandas", 
    "ipyleaflet", 
    "plotly",
    "lorem_text", 
    "matplotlib",
    "psycopg2-binary",
    "scipy", 
    "pandas",
    "networkx",
    "openpyxl",
    "rasterio",
    "boto3",
    "cryptography",
    "ipydatagrid",
    "requests_oauthlib",
    "python-dotenv",
]

[tool.hatch.version]
path = "tomorrowcities/__init__.py"



[project.urls]
Home = "https://github.com/TomorrowsCities/tomorrowcities"

[tool.black]
line-length = 160

[tool.isort]
profile = "black"
